WIP: Getting the Models worked out
Posts do not have content, title, slug, or really much of anything on
their own.  They hold together the concept that a Post is actually a
series of revisions with some level of visibility state and that
Urls (mostly the slug part) also changes with time but always points
to the same concept.  Still need to deal with the idea that there is
only one canonical Url for a Post (the others will 301)