Using DotEnv to manage INFO setup rather than yaml file
I had originally thought of moving this to the Database and just
having the user be redirected to a form to fill in this data on first
visit similar to the way the the upstream Obtvse handles users (you
are propmpted to create a user and then thereafter are required to
login) after thinking about it some more I decided to use dotenv for
development rather than a home grown yaml/erb solution and that any
modern PaaS that this would be deployed to is going to make it super
easy to manage the ENV variables.  Thus a README for deployment should
be sufficient.

I also pulled out a bunch of "dead" code.